【修改描述】cleancode 【DTS/CR/AR单号】 #0 【合入版本】BeiMing 24.2.T1.B001
# message auto-generated for no-merge-commit merge:
merge dev-cleancode into master

【微重构】cleancode


Match-id-204bd7f3c43d384ed92c172c78819bad2c7a7a9b

+++ src/config/bio_config_instance.cpp
@@ -212,7 +212,7 @@ BResult BioConfig::AutoConfigDaemonLogAndOther(const ConfigurationPtr &conf)
     mDaemonConfig.enableCli = conf->GetStr(BIO_CLI_TOOLS_ENABLE.first) == "true";
     mDaemonConfig.enablePrometheus = conf->GetStr(PROMETHEUS_ENABLE.first) == "true";
     mDaemonConfig.listenAddress = conf->GetStr(PROMETHEUS_LISTEN_ADDRESS.first);
-    mDaemonConfig.scrapeIntervalSec = conf->GetInt(PROMETHEUS_SCRAPE_INTERVAL_SEC.first);
+    mDaemonConfig.scrapeIntervalSec = static_cast<uint32_t>(conf->GetInt(PROMETHEUS_SCRAPE_INTERVAL_SEC.first));
 
     std::string scene = conf->GetStr(WORK_SCENE.first);
     if (scene == "none") {
